fix: changed pattern of creating preferences
