+++ packages/runtime/src/index.ts
@@ -1,2 +1,3 @@
 export * from './microapp-runtime';
 export * from './microapp';
+export * from './constants';
