--- packages/user-preferences/src/constants.ts
@@ -1,1 +1,0 @@
-export const PRODUCTION_MARKETPLACE_HOST_URL = 'https://www.microapp.io/';
