+++ packages/user-preferences/src/index.ts
@@ -1,3 +1,1 @@
-export * from './create-user-preferences';
-export * from './sandbox-user-preferences';
 export * from './user-preferences';
