+++ packages/user-preferences/src/preferences-repo.ts
@@ -1,0 +1,6 @@
+import { UserPreferencesData } from './types';
+
+export interface PreferencesRepo {
+  getPreferences(): UserPreferencesData;
+  onUpdate(callback: (data?: UserPreferencesData) => void): () => void;
+}
