+++ packages/user-preferences/src/types.ts
@@ -1,0 +1,10 @@
+export type UserPreferencesData = {
+  theme?: string;
+  lang?: string;
+};
+
+export type UserPreferencesOptions = {
+  sandbox?: UserPreferencesData;
+};
+
+export type SandboxPreferencesOptions = UserPreferencesData;
